Refactor for a smaller routes file
