+++ Sources/App/CircleCI/Models/CircleCIBuildStepAction.swift
@@ -1,0 +1,9 @@
+import Vapor
+
+public struct CircleCIBuildStepAction: Content {
+    public let outputURL: URL?
+    
+    enum CodingKeys: String, CodingKey {
+        case outputURL = "output_url"
+    }
+}
