+++ Sources/App/CircleCI/Models/CircleCIPullRequest.swift
@@ -1,0 +1,12 @@
+import Vapor
+
+
+public struct CircleCIPullRequest: Content {
+    let headCommit: String
+    let url: URL
+    
+    enum CodingKeys: String, CodingKey {
+        case headCommit = "head_sha"
+        case url
+    }
+}
