+++ Sources/App/Models/PerformanceTestResults.swift
@@ -1,14 +1,14 @@
 import Foundation
 
-struct TestResults: Codable {
+public struct PerformanceTestResults: Codable {
     public let name: String
     public let expected: Double
     public let average: Double
     public let change: String
 }
 
-extension TestResults: Rowable {
-    func fields() -> [String] {
+extension PerformanceTestResults: Rowable {
+    public func fields() -> [String] {
         return [name, String(expected), String(average), change]
     }
 }
